feat: TransactionManager 추상화
